Don't use int16_t for VITS level

+++ vits.h
@@ -27,7 +27,7 @@ typedef struct {
 	int16_t *line[4];
 } vits_t;
 
-extern int vits_init(vits_t *s, unsigned int sample_rate, int width, int lines, int16_t level);
+extern int vits_init(vits_t *s, unsigned int sample_rate, int width, int lines, int level);
 extern void vits_free(vits_t *s);
 
 extern int vits_render(vid_t *s, void *arg, int nlines, vid_line_t **lines);
